Refactor authorization logic to improve client validation and error handling
