+++ package.json
@@ -26,7 +26,7 @@
     "eslint-plugin-promise": "^7.2.1",
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-security": "^3.0.1",
-    "eslint-plugin-sonarjs": "^3.0.4",
+    "eslint-plugin-sonarjs": "^3.0.5",
     "prettier": "^3.6.2",
     "vitest": "^3.2.4",
     "@playwright/test": "^1.55.0"
